move the ui focused files to ui package
