+++ src/main/java/iris/Main.java
@@ -2,6 +2,7 @@ package iris;
 
 import java.io.IOException;
 
+import iris.ui.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
